requirements.txt
alembic
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
python-jose
pydantic[email]
passlib[bcrypt]
pytz
psycopg2-binary
fastapi-pagination
openpyxl
requests
apscheduler
pydantic
pydantic-settings
python-dotenv
pytest
httpx